Version 15.192.2

+++ swad_image.c
@@ -450,15 +450,18 @@ void Img_RemoveImageFile (const char *ImageName)
   {
    char FullPathImgPriv[PATH_MAX+1];
 
-   /***** Build path to private file *****/
-   sprintf (FullPathImgPriv,"%s/%s/%c%c/%s.jpg",
-	    Cfg_PATH_SWAD_PRIVATE,Cfg_FOLDER_IMG,
-	    ImageName[0],
-	    ImageName[1],
-	    ImageName);
+   if (ImageName[0])
+     {
+      /***** Build path to private file *****/
+      sprintf (FullPathImgPriv,"%s/%s/%c%c/%s.jpg",
+	       Cfg_PATH_SWAD_PRIVATE,Cfg_FOLDER_IMG,
+	       ImageName[0],
+	       ImageName[1],
+	       ImageName);
 
-   /***** Remove private file *****/
-   unlink (FullPathImgPriv);
+      /***** Remove private file *****/
+      unlink (FullPathImgPriv);
 
-   // Public links are removed automatically after a period
+      // Public links are removed automatically after a period
+     }
   }
